requirements.txt
fastapi
uvicorn
aiohttp
fastapi_utils
python-dotenv